Add pagination to acitities feed.
This adds :next_page and :previous_page methods to FitnessActitiesFeed,
which return new instances. The existing :next and :previous accessors
work as before, although they're not very useful now that pagination
is available.

For example, to fetch every item ever posted:

  feed = @user.fitness_activities
  items = []

  while feed
    items += feed.items
    feed = feed.next_page
  end

This would obviously be a very bad idea in production.